blog/requirements.txt
bootstrap-admin==0.4.1
bs4==0.0.1
Django==2.2.1
django-haystack==2.8.1
django-simple-captcha==0.5.9
django-tinymce==2.8.0
jieba==0.39
mysqlclient==1.4.2
oss2==2.6.1
Pillow==5.4.1
requests==2.21.0
xadmin==2.0.1
